Plumbing cleanup around modulation switcher
- Remove nine unused findXxxChainIndex() private methods from
  ChainListPanel. All nine had zero call sites; the codebase uses
  effectChain.getEffectByName() directly everywhere.
- Extract activeModulationEffectName() helper to replace four copies
  of the same five-way ternary computing the active engine's name.
  Adding a sixth modulation engine in future now touches one site
  instead of four.
- Use getInt instead of getDouble for the integer "stages" (Phaser)
  and "mode" (Tremolo) preset fields. Behaviourally identical, matches
  the pattern used by every other int field in PresetManager.
- Add null guard on oversampler in Tremolo::processBias defensively;
  prepare() is contractually called before process() but this is
  cheap belt-and-suspenders matching the existing guards in reset()
  and getLatencySamples().

+++ src/preset/PresetManager.cpp
@@ -623,7 +623,7 @@ void PresetManager::applyPreset(const Preset& preset)
                 phaser->setDepth(getDouble(v, "depth", 0.5));
                 phaser->setFeedback(getDouble(v, "feedback", 0.3));
                 phaser->setMix(getDouble(v, "mix", 0.5));
-                phaser->setStages(static_cast<int>(getDouble(v, "stages", 0)));
+                phaser->setStages(getInt(v, "stages", 0));
             }
         }
     }
@@ -653,7 +653,7 @@ void PresetManager::applyPreset(const Preset& preset)
                 tremolo->setBypassed(getBool(v, "bypassed", true));
                 tremolo->setRate(getDouble(v, "rate", 0.4));
                 tremolo->setDepth(getDouble(v, "depth", 0.5));
-                tremolo->setMode(static_cast<int>(getDouble(v, "mode", 0)));
+                tremolo->setMode(getInt(v, "mode", 0));
                 tremolo->setWidth(getDouble(v, "width", 0.0));
                 tremolo->setOutput(getDouble(v, "output", 0.5));
             }
